프로필, 프로필 업데이트 페이지 완료
프로필 페이지와 프로필 수정 페이지가 다름
  - django에서 프로필 수정관련 코드 분리
프로필 수정시 url의 username params로 다른 사람의 프로필 수정 가능성
존재
  - views.py에서 로그인한 user와 다르면 400error 발생시키도록함
  - vue에서 네비게이션 가드 할 예정

남은 일
  - 팔로우 구현

+++ server/accounts/views.py
@@ -1,13 +1,15 @@
-from django.shortcuts import get_object_or_404, get_list_or_404
-from django.contrib.auth import get_user_model
-
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
+
+from django.shortcuts import get_object_or_404, get_list_or_404
+from django.contrib.auth import get_user_model
+
 from .serializers import UserSerializer
 
 # Create your views here.
 
+# 회원가입
 @api_view(['POST'])
 def signup(request):
     serializer = UserSerializer(data=request.data)
@@ -15,9 +17,10 @@ def signup(request):
     password_confirm = request.data.get('password_confirm')
     nickname = request.data.get('nickname')
 
+    # 비밀번호 검증
     if password != password_confirm:
-        data = {'password_unmatch': '비밀번호가 일치하지 않습니다.',}
-        return Response(data, status=status.HTTP_400_BAD_REQUEST)
+        context = {'password_unmatch': '비밀번호가 일치하지 않습니다.',}
+        return Response(context, status=status.HTTP_400_BAD_REQUEST)
 
     if serializer.is_valid(raise_exception=True):
         user = serializer.save()
@@ -27,9 +30,10 @@ def signup(request):
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
-@api_view(['GET', 'PUT', 'DELETE'])
+@api_view(['GET', 'DELETE'])
 def profile(request, username):
     profile_user = get_object_or_404(get_user_model(), username=username)
+    # 프로필 조회
     if request.method == 'GET':
         context = {
             'username': profile_user.username,
@@ -37,7 +41,59 @@ def profile(request, username):
             'introduce': profile_user.introduce,
         }
         return Response(context)
-    elif request.method == 'PUT':
-        pass
-    elif request.method == 'DELETE':
-        profile_user.delete()
+    
+    # 계정 삭제
+    elif request.data.get('username') == request.user.username:
+        if request.method == 'DELETE':
+            profile_user.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+    else:
+        context = {
+            'faild': '실패'
+        }
+    return Response(context, status=status.HTTP_400_BAD_REQUEST)
+
+
+@api_view(['GET', 'PUT'])
+def update(request, username):
+    if request.user.username == username:
+        user = get_object_or_404(get_user_model(), username=username)
+        # 프로필 수정 페이지 조회
+        if request.method == 'GET':
+            context = {
+                'username': user.username,
+                'nickname': user.nickname,
+                'introduce': user.introduce,
+            }
+            return Response(context)
+        
+        # 프로필 수정
+        elif request.method == 'PUT':
+            password = request.data.get('password')
+            password_confirm = request.data.get('password_confirm')
+            nickname = request.data.get('nickname')
+            introduce = request.data.get('introduce')
+
+            # 비밀번호 검증
+            if password and password != password_confirm:
+                context = {'password_unmatch': '비밀번호가 일치하지 않습니다.',}
+                return Response(context, status=status.HTTP_400_BAD_REQUEST)
+
+            # 프로필 수정
+            else:
+                if password:
+                    user.set_password(password)
+                user.nickname = nickname if nickname else user.nickname
+                user.introduce = introduce if introduce else user.introduce
+                user.save()
+                context = {'success': '수정이 완료되었습니다.'}
+                return Response(status=status.HTTP_204_NO_CONTENT)
+    else:
+        message = '자신의 프로필만 수정할 수 있습니다.'
+    context ={'message': message}
+    return Response(context, status=status.HTTP_400_BAD_REQUEST)
+
+
+@api_view(['GET', 'POST'])
+def follow(request):
+    pass
